v1/requirements.txt
# Core dependencies
notion-client==2.2.1
requests==2.31.0
beautifulsoup4==4.12.3
loguru==0.7.2
python-dotenv==1.0.0

# Optional dependencies
lxml==5.1.0  # For faster HTML parsing